Fix translation system

+++ make_translations.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-LANGUAGES="en es ca@valencia"
+LANGUAGES="en es_ES ca_ES"
 GETTEXT_SCRIPT="pygettext.py"
 MSGFMT_SCRIPT="msgfmt.py"
 OUTDIR="locales"
@@ -9,6 +9,7 @@ SEARCH_TRANSLATABLE_FILES_PATH="."
 PYVERSION="$(python -c "import sys; print(str(sys.version_info.major)+'.'+str(sys.version_info.minor))" )"
 XGETTEXT="$(find /usr/ -name "${GETTEXT_SCRIPT}" 2> /dev/null |grep ${PYVERSION})"
 MSGFMT="$(find /usr/ -name "${MSGFMT_SCRIPT}" 2> /dev/null |grep ${PYVERSION})"
+MSGMERGE="$(which msgmerge)"
 
 TRANSLATABLE_FILES="$(find "${SEARCH_TRANSLATABLE_FILES_PATH}" -type f | while read in; do if file -i "${in}" |grep -q x-python; then echo "${in}"; fi; done; )"
 POTFILE="./${OUTDIR}/${DOMAIN}.pot"
@@ -26,10 +27,14 @@ $XGETTEXT -d $DOMAIN -o ${POTFILE} ${TRANSLATABLE_FILES}
 echo ${POTFILE} generated!
 
 for lang in $LANGUAGES; do
-    copydir="${OUTDIR}/${lang}"
+    copydir="${OUTDIR}/${lang}/LC_MESSAGES"
     pofile="${copydir}/${DOMAIN}.po"
     mofile="${copydir}/${DOMAIN}.mo"
     mkdir -p ${copydir}
-    cp ${POTFILE} ${pofile}
+    if [ -f "${pofile}" ]; then
+	${MSGMERGE} -U ${pofile} ${POTFILE}
+    else
+        cp ${POTFILE} ${pofile}
+    fi
     ${MSGFMT} -o ${mofile} ${pofile}
 done
